+++ src/ar/sceneviewer/__test__/SceneViewer.test.ts
@@ -1,6 +1,6 @@
 import { Box3, Color, Euler, Mesh, Object3D, Vector3 } from 'three';
 import { DIVEScene } from '../../../scene/Scene';
-import { ARSystemOptions } from '../../AR';
+import { ARSystemOptions } from '../../ARSystem';
 import { SceneViewer } from '../SceneViewer';
 import { SystemInfo } from '../../../info/Info';
